+++ backend/src/main/java/ru/mephi/trainer/rest/dto/response/ProfileResponse.java
@@ -9,7 +9,7 @@ public class ProfileResponse {
     private String email;
     private String firstName;
     private String lastName;
-    private Double totalScore;
+    private Integer totalScore;
     private List<TrainerProgressPercentResponse> trainerProgressPercent;
     private String createdAt;
 }
